fix: render job experiences dynamically
Previously, a lot of time was wasted by copying HTML elements and
changing the text whenever I needed to update the file. When HTML tags
were not closed (as with the case when there were many tags), it was a
pain to debug the exact tag that was unclosed.

This commit fixes it by dynamically creating HTML template strings and
injecting it into specific HTML nodes.